Move cache into /utilities
